requirements.txt
Flask==3.1.0
PyPDF2==3.0.1
pdfplumber==0.11.5
camelot-py==1.0.0
pdfminer.six==20240706
tika==2.6.0
pytesseract==0.3.13
pdf2image==1.17.0